refactor: rename web-dev-server -> webpack-dev-server
