+++ src/@rocket-scripts/webpack-dev-server/utils/patchProxyLogger.ts
